[PATCH] libpayload:  Fix overflow in _delay function
On faster machines, delta might be more then 32 bits

Acked-by: Patrick Georgi <patrick.georgi@coresystems.de>


git-svn-id: svn://svn.coreboot.org/coreboot/trunk@3677 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ payloads/libpayload/i386/timer.c
@@ -75,7 +75,7 @@ unsigned int get_cpu_speed(void)
 	return cpu_khz;
 }
 
-static inline void _delay(unsigned int delta)
+static inline void _delay(unsigned long long delta)
 {
 	unsigned long long timeout = rdtsc() + delta;
 	while (rdtsc() < timeout) ;
